requirements.txt
notebook
ipywidgets
tqdm
python-dotenv
openai
streamlit
rembg
mlflow
asyncio
aiohttp
-e .